+++ integration/simple-esmodule-interop/jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  testMatch: ['<rootDir>/build/**/*-test.js'],
+  testEnvironment: 'node',
+};