+++ integration/simple-proto2/simple.ts
@@ -1,6 +1,6 @@
 /* eslint-disable */
-import { util, configure, Writer, Reader } from 'protobufjs/minimal';
 import * as Long from 'long';
+import * as _m0 from 'protobufjs/minimal';
 
 export const protobufPackage = 'simple';
 
@@ -45,15 +45,15 @@ function createBaseIssue56(): Issue56 {
 }
 
 export const Issue56 = {
-  encode(message: Issue56, writer: Writer = Writer.create()): Writer {
+  encode(message: Issue56, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
     if (message.test !== 1) {
       writer.uint32(8).int32(message.test);
     }
     return writer;
   },
 
-  decode(input: Reader | Uint8Array, length?: number): Issue56 {
-    const reader = input instanceof Reader ? input : new Reader(input);
+  decode(input: _m0.Reader | Uint8Array, length?: number): Issue56 {
+    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
     let end = length === undefined ? reader.len : reader.pos + length;
     const message = createBaseIssue56();
     while (reader.pos < end) {
@@ -108,9 +108,9 @@ export type Exact<P, I extends P> = P extends Builtin
 
 // If you get a compile-error about 'Constructor<Long> and ... have no overlap',
 // add '--ts_proto_opt=esModuleInterop=true' as a flag when calling 'protoc'.
-if (util.Long !== Long) {
-  util.Long = Long as any;
-  configure();
+if (_m0.util.Long !== Long) {
+  _m0.util.Long = Long as any;
+  _m0.configure();
 }
 
 function isSet(value: any): boolean {